+++ CHANGELOG.md
@@ -20,3 +20,4 @@ All notable changes to this project will be documented in this file.
 ### Removed
 
 - Removed the skipCache client option in operations that do not read cache.
+- Removed unimplemented `DeleteAllOf()` functionality.